[Cases] Refactor integration helper functions (#149640)
This PR refactors the cases integration test API helper functions so
that splitting up the `utils.ts` file will be easier in the future
because it will all be hidden under the `api/index.ts` import. That way
moving functions out of the `index.ts` file into their own will not
require a ton of test files to be updated.