[ML] Adjusts anomaly detection screenshot script (#150584)
## Summary

This PR makes adjustments on one of the anomaly detection screenshot
scripts:
* renames the test file to `influencers` from `population_analysis`,
* renames one of the screenshots in the script,
* removes two screenshot creation steps.

The reason behind all of this is that the doc page about population
analysis has been replaced by another, higher-level page that doesn't
contain any screenshots. Two screenshots covered in this test file
became unnecessary because of this. The third one, however, captures
every information required for another screenshot in the docs about
influencers.