[build] Fix chromium install path (#151208)
Fixes an issue introduced in 898048edabb071b81a02c45144e66f117be9c4c6

Currently during builds chromium is installed to the `x-pack/reporting`
directory. Under the new package management system plugins are installed
in node_modules, and the reporting plugin is not able to find the path
of chromium. This patches the path.

Ideally we could install chromium before the plugin is installed as a
package, but plugin installation is run once for all platforms, and
later reporting is run for each platform.

https://buildkite.com/elastic/kibana-artifacts-snapshot/builds/1222