Narrow scope of forced minimatch resolution (#151290)
The "global" forced resolution of the `minimatch` package
unintentionally downgraded some v5.x dependencies to v3.x. The only
package where we need to force upgrade `minimatch` is the `globule`
package which depends on `minimatch@~3.0.2`. So in this PR I've scoped
the resolution to only concern itself with the `globule` dependencies.