[Fleet] Hardcode log levels in agent drodpdown (#151451)
Fixes https://github.com/elastic/kibana/issues/151177

## Summary

In the log levels list we were using the `autocomplete` service to get
the available logs, but sometimes the list wouldn't appear. In
https://github.com/elastic/kibana/pull/151030 I refactored that code to
use the latest version of the API, but it's a better solution to just
hard code this list since the possible levels are very few.

<img width="851" alt="Screenshot 2023-02-16 at 15 35 53"
src="https://user-images.githubusercontent.com/16084106/219394351-a73b24ec-c550-42fc-8d4c-742f1c0091e7.png">

With this change the log levels should always be visible as in the
picture.