[Fleet] Fixing update tags status reporting complete too early (#151330)
## Summary

Fixes https://github.com/elastic/kibana/issues/150654

Found a bug in update tags, that when writing out action results, the
count was not correct in case of conflicts (less agents were updated
than agentIds).
This resulted in action result count reaching the total early, and
marking the action status complete.

This caused a failure in performance tests, because the action was set
to complete, but actually not all agents had the new tag yet.

With the fix on action results I see the ack count correct, and the
action moving to COMPLETE when all retries have finished.

Added another fix for `nbAgentsActionCreated` incorrectly showing higher
count than expected. This value comes from the sum of agentIds in the
`.fleet-actions` doc. Changed the implementation so that
`.fleet-actions` is updated on every retry to only record agents
updated, failed, and on last retry conflicted. This ensures that the
`nbAgentsActionCreated` will be accurate.

Tested with 50k agents, see recording here
https://github.com/elastic/kibana/pull/151330#discussion_r1109526053

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios