[Infrastructure UI] Remove 'Introducing a new Hosts analysis experience' after clicked on once (#151433)
Closes [#151411](https://github.com/elastic/kibana/issues/151411)
[Flaky Test fix]
Closes https://github.com/elastic/kibana/issues/150143
Closes https://github.com/elastic/kibana/issues/150144
Closes https://github.com/elastic/kibana/issues/150145
Closes https://github.com/elastic/kibana/issues/150146

## Summary
This PR fixes the hosts view link on the inventory - after the user
clicks on the link once it won't be visible anymore during the same
browser session. Also, there was a missing step in one of the test
helpers so I added it and now the tests should not be flaky anymore.

## Testing 


https://user-images.githubusercontent.com/14139027/220119166-28081ae7-0d08-4a5b-815f-21268157cfdd.mov


Go to inventory and click hosts try it badge/link. 
The hosts view page should be displayed. 
Go back to inventory - the link should not be visible.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>