[ML] Transforms: Fix to retain existing query when there's no base filter criteria. (#151665)
When a data view had a time field, `getPreviewTransformRequestBody()`
would even add an empty array of base filter criteria. The code later on
would then not recognize that as a default query and add an empty bool
filter query. This could even end up as nested empty bool filters when
cloning a transform. This fixes it by adding a check for an empty array of base filter
criteria. A jest unit test was also added to cover the original problem.