[Profiling] Normalize samples by time for differential flamegraph (#152158)
This PR appropriately scales the samples in the differential
flamegraph's tooltip when time-normalized mode is selected.

Fixes https://github.com/elastic/prodfiler/issues/3038

### Notes

* Respective values for the normalization menu and differential
flamegraph are now defined in the parent view so that both elements
remain in sync. Previously only the normalization menu had the scaling
factors for time-normalized mode, thus, the tooltip in the differential
flamegraph was not accurate.
* The prior scaling factors for scale-normalized mode are remembered as
long as a user is on the differential flamegraph. Thus, a user can
update the time ranges, the format (Abs vs Rel), and the normalization
mode without losing their previously chosen scale-normalized values.
* The time-normalized scaling factors continue to remain immutable.
* Due to artifacts related to floating-point division, the adjusted
samples may not be whole integers.