Replacing the alert icon for an error icon in erroneous situations (#151413)
## Summary

This PR is aiming to replace the `alert` icon that we use in error
states, with the `error` icon. It's important to differentiate them
clearly as errors and warning messages are semantically different
states. It's widely considered that errors maybe fatal errors and
warnings are never fatal.

You should receive an error message when the issue is terminal, this
doesn't always mean that the operation stops completely, but the task is
not complete.

You should receive a warning message when the program detects that
something is not behaving right, but it didn't cause any termination.


Before / After

![image](https://user-images.githubusercontent.com/13353203/219329971-50c95178-6fdd-48d6-b4b5-43e648e6fb9a.png)

Example of how it looks in the Profiling UI, when two errors appear:
<img width="1726" alt="image"
src="https://user-images.githubusercontent.com/13353203/219330082-6f638ffb-995b-40eb-88d7-cfc6026e006d.png">

![image](https://user-images.githubusercontent.com/13353203/219333118-1876b085-f7dc-46ed-bbea-e75675c928ad.png)

## Related to
https://github.com/elastic/eui/issues/6554

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>