[Response Ops][Actions] Update json editor value for default body (#152133)
## Summary

See https://github.com/elastic/kibana/pull/151851 for link to bug
description and details about the fix. This PR adds a fix for behavior
in the ES index connector params component that appeared as a result of
the changes in the JSON editor. When the JSON editor is modified to
re-render with `useEffect`, the ES index connector `onBlur` behavior
would auto-populate the editor with `[object Object]`. This PR updates
the component to work with the re-render.

---------

Co-authored-by: semd <sergi.massaneda@elastic.co>
Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>