[Kibana migrator test kit] Flush logger after running migrations (#152818)
This PR addresses flakiness of:

- https://github.com/elastic/kibana/issues/152472
- https://github.com/elastic/kibana/issues/152448

These tests fail because we have a race condition. We are checking
certain conditions in the log files, and sometimes we do it before the
logs are actually written.

With the Kibana migrator test kit, we can actually flush + wait for all
the logging appenders. This way, we are sure that the logs are complete
when we inspect them in the tests.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>