[ML] AIOps: Fix use full data button when time range returns null. (#152889)
In certain cases the endpoint to get the full date range may return `{
start: null, end: null }`, for example when querying against a
frozen-only index but with the option enabled to exclude the frozen
tier. In AIOps we added a callback to push that information to the URL's
global state but that would corrupt then the global state with the page
crashing.

This PR fixes it with an update to `setFullTimeRange`. The function will
now only return the time range if it's fully populated. After that, the
`FullTimeRangeSelector` component will only call the callback if it
receives such a time range and will skip the callback if
`setFullTimeRange` returned `undefined`.

Jest unit tests for `setFullTimeRange` and `FullTimeRangeSelector` have
been updated/created to cover both cases with populated time ranges and
without.