[Enterprise Search] Support pages without ent-search (#152969)
## Summary

Introduces new config properties related to running `enterprise_search`
plugin without the `ent-search` service running. All three default to
`true` and retain the current behavior when they are set to `true`.

`canDeployEntSearch` - Setting to false will result in not showing the
Setup Guide for content pages when `config.host` is not set. Essentially
ungating indices, engines etc. pages.
`hasConnectors` - setting to `false` will result in hiding the connector
ingestion method when creating a new index.
`hasDefaultIngestPipeline` - setting to `false` will result in hiding
the index Pipelines and Content Settings page, until we can migrate
management of default ingest pipeline to the ES module
`hasNativeConnectors` - setting to `false` will result in hiding the
native connector ingestion method when creating a new index.
`hasWebCrawler` - setting to `false` will result in hiding the web
crawler ingestion method when creating a new index.

`hasNativeConnectors` & `hasWebCrawler` are intended to be temporary and
can be removed once we are sure the connectors service and web crawler
will be available in serverless.