[Enterprise Search]: Engines - parse field capabilities (#152953)
## Summary

Original: #152248

### TODO

- ~~[ ] Decide if more is needed for multi-fields~~
- They currently show up in the `fields` list just like object & nested
sub-fields but they need to be rendered slightly differently. So not
sure if we need to add something to the object to differentiate they are
multi-fields. Which would require writing something that can determine
that difference, which could just be `type !== 'object' && type !==
'nested'` but I'm not :100: on that.
- i think we can make these changes if needed when implementing the UI
- [x] Confirm if we want to include metadata fields, currently filtering
them out of the list

these tasks will be completed in a separate PR/issue
- ~~[ ] Use `fields` in the UI instead of `field_capabilities`~~
- ~~[ ] remove `field_capabilities` from API response~~

### Checklist

Delete any items that are not applicable to this PR.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

Co-authored-by: Rodney Norris <rodney.norris@elastic.co>