github-action: support deploy my kibana instance using the Observability test clusters (#153106)
## Summary

As we discussed in an email, this is our proposal to facilitating the
Observability teams to create Kibana instances attached to the
Observability test environments by using the GitHub command
`/oblt-deploy`

If an Elastician added a GitHub comment then, the automation will create
the Kibana instance based on the PR changes. Then a comment with the
link to the GitHub issue that contains all the configuration details.
The GitHub issue is not accessible to the public since it's under a
private GitHub repository.