[ML] Fix apiDoc script  (#153366)
## Summary

Previously, apidoc scripts were transpiled to TS and the JS files were
passed as parameters to the npm command. With the recent Kibana changes
related to `tsconfig`, the output structure has changed and caused
`apiDocs` to fail.

With this PR I introduced a scripts folder that utilizes
[setup_node_env](https://github.com/elastic/kibana/blob/b701e1e152e43c9739ebc8675c5ada4ffa700dd1/src/setup_node_env/index.js#L10)
and allows us to omit `tsc` step.

Also, the `getTypeAtLocation` method started throwing an error in some
cases. For now, I've added an error handling and need to investigate
what's off.