Fix flaky RBAC Legacy test due to archive containing tasks and rules together (#153803)
Resolves https://github.com/elastic/kibana/issues/153800
Resolves https://github.com/elastic/kibana/issues/142704
Resolves https://github.com/elastic/kibana/issues/153801
Resolves https://github.com/elastic/kibana/issues/142947
Resolves https://github.com/elastic/kibana/issues/140867

Similar to
https://github.com/elastic/kibana/pull/152841#discussion_r1147716618,
the rule and tasks archives don't seem to play nicely when combined. The
flakiness goes away when loading the rules then the tasks in sequence.
Otherwise, the tasks sometimes run before it can find the rule, causing
the task to delete itself.

I took a look at why the task would run an not be able to find the rule.
My best guess after looking at a failing flaky test is that the task
manager migration completes before the .kibana. And while .kibana
migrates, the task runs and fails to load the task because the .kibana
index is in an interim state.

Flaky test runner:
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/2045

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>