[Watcher] Remove deprecated savedObjectsClient (#154172)
## Summary
Fixes https://github.com/elastic/kibana/issues/154035

This PR removes the saved objects client from the client-side code of
the Watcher plugin. It seems not to be used anywhere so the removal
should not cause any changes to the UI.
To test this PR, it should be enough to create/update/delete a couple of
simple and advanced watches.