[Cases] Filter out alerts already attached to the case on the backend (#154629)
## Summary

This PR filters out alerts already attached to a case on the backend.
The schema of the alerts in a case is:

```
{
  alertId: string | string[]
  index: string | string[]
}
```

So the `alertId` can contain multiple alert ids. In this case, if an ID
is attached to the case it will be removed from the `alertId` array. If
all alerts are attach to the case no attachment will be created.

### Checklist

Delete any items that are not applicable to this PR.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

### For maintainers

- [x] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)